requirements.txt
asgiref==3.8.1
Django==5.2
django-timescaledb==0.2.13
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3

requests~=2.32.3
channels~=4.2.2
python-dotenv~=1.1.0
dotenv~=0.9.9
pytz~=2025.2